@@ -29,39 +29,46 @@
   },
   "private": true,
   "dependencies": {
-    "tslib": "^2.0.0",
-    "@storybook/builder-webpack5": "~6.4.5",
-    "@storybook/manager-webpack5": "~6.4.5"
+    "@storybook/builder-webpack5": "6.4.22",
+    "@storybook/core-server": "~6.4.12",
+    "@storybook/manager-webpack5": "6.4.22",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
+    "react-is": "18.0.0",
+    "tslib": "^2.0.0"
   },
   "devDependencies": {
     "@nrwl/cli": "13.3.0",
-    "@nrwl/devkit": "13.3.0",
-    "@nrwl/eslint-plugin-nx": "13.3.0",
-    "@nrwl/express": "13.3.0",
-    "@nrwl/jest": "13.3.0",
-    "@nrwl/linter": "13.3.0",
-    "@nrwl/nest": "13.3.0",
-    "@nrwl/node": "13.3.0",
-    "@nrwl/nx-plugin": "13.3.0",
-    "@nrwl/react": "13.3.0",
-    "@nrwl/storybook": "13.3.0",
-    "@nrwl/tao": "13.3.0",
-    "@nrwl/workspace": "13.3.0",
+    "@nrwl/devkit": "14.3.0",
+    "@nrwl/eslint-plugin-nx": "14.3.0",
+    "@nrwl/express": "14.3.0",
+    "@nrwl/jest": "14.3.0",
+    "@nrwl/js": "14.3.0",
+    "@nrwl/linter": "14.3.0",
+    "@nrwl/nest": "14.3.0",
+    "@nrwl/node": "14.3.0",
+    "@nrwl/nx-plugin": "14.3.0",
+    "@nrwl/react": "14.3.0",
+    "@nrwl/storybook": "14.3.0",
+    "@nrwl/workspace": "14.3.0",
     "@phenomnomnominal/tsquery": "^4.1.1",
-    "@types/jest": "27.0.2",
-    "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "5.3.1",
-    "@typescript-eslint/parser": "5.3.1",
+    "@swc-node/register": "^1.4.2",
+    "@swc/core": "^1.2.173",
+    "@types/jest": "27.4.1",
+    "@types/node": "17.0.41",
+    "@typescript-eslint/eslint-plugin": "5.24.0",
+    "@typescript-eslint/parser": "5.24.0",
     "dotenv": "10.0.0",
-    "eslint": "8.2.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
-    "jest": "27.2.3",
+    "jest": "27.5.1",
     "node-fetch": "^2.6.1",
-    "prettier": "2.3.2",
-    "ts-jest": "27.0.5",
-    "ts-node": "~9.1.1",
+    "nx": "14.3.0",
+    "prettier": "2.6.2",
+    "react-test-renderer": "18.0.0",
+    "ts-jest": "27.1.4",
+    "ts-node": "10.8.1",
     "tslib": "^2.0.0",
-    "typescript": "4.3.5"
+    "typescript": "4.7.3"
   }
 }
-
